// src/modules/organization-settings/entities/rbac.enums.ts
/**
 * RBAC Enums
 * Centralized enum definitions to avoid circular dependencies
 */

export enum RoleCategory {
  SYSTEM_ADMIN = 'SYSTEM_ADMIN',
  TENANT_ADMIN = 'TENANT_ADMIN',
  HR_ADMIN = 'HR_ADMIN',
  HR_MANAGER = 'HR_MANAGER',
  LINE_MANAGER = 'LINE_MANAGER',
  PAYROLL_ADMIN = 'PAYROLL_ADMIN',
  FINANCE_VIEWER = 'FINANCE_VIEWER',
  EMPLOYEE_SELF_SERVICE = 'EMPLOYEE_SELF_SERVICE',
  AUDITOR = 'AUDITOR',
  RECRUITER = 'RECRUITER',
  COMPLIANCE_OFFICER = 'COMPLIANCE_OFFICER',
  DEPARTMENT_HEAD = 'DEPARTMENT_HEAD',
}

export enum DataScopeLevel {
  OWN_DATA_ONLY = 'OWN_DATA_ONLY',
  DIRECT_REPORTS = 'DIRECT_REPORTS',
  ORGANIZATION_HIERARCHY = 'ORGANIZATION_HIERARCHY',
  DEPARTMENT = 'DEPARTMENT',
  DIVISION = 'DIVISION',
  BUSINESS_UNIT = 'BUSINESS_UNIT',
  LEGAL_ENTITY = 'LEGAL_ENTITY',
  COUNTRY = 'COUNTRY',
  REGION = 'REGION',
  GLOBAL = 'GLOBAL',
}